# contrat-openapi.yaml
openapi: "3.0.3"
servers:
  - url: "http://localhost:3000/v1"
    description: "Serveur de développement local"
  - url: "https://staging-api.thermosense.com/v1"
    description: "Serveur de staging pour tests"
  - url: "https://api.thermosense.com/v1"
    description: "Serveur de production"
info:
  title: ThermoSense API
  version: "1.0.0"
  description: |
    API de gestion de la plateforme ThermoSense —
    supervision de capteurs et pilotage d'actionneurs pour le confort thermique des bâtiments.
    Le controle d'acces combine des scopes JWT, des restrictions par role et des verifications BOLA sur la zone ou la ressource cible.

    ## Versioning et politique d'évolution

    **Stratégie :** versioning par **préfixe d'URI** (`/v1`, `/v2`, ...). Tous les environnements
    (local, staging, production) servent la version courante sous `/v1`. Le numéro majeur
    de l'URI est synchronisé avec le `version` SemVer ci-dessus.

    **Compatibilité ascendante au sein d'une majeure (`/v1`)** : les évolutions suivantes
    sont considérées **non-breaking** et peuvent être publiées sans nouvelle version :
    - Ajout d'un endpoint, d'un champ optionnel en réponse, d'un nouvel `enum` toléré
      par le client (les clients doivent ignorer les champs inconnus).
    - Ajout d'un paramètre de requête optionnel.
    - Ajout d'un nouveau code HTTP 2xx documenté à l'avance.

    Sont considérées **breaking** et nécessitent `/v2` :
...

...
    chaque réponse concernée.

...

...

...

...
    |---|---|---|
...

...

...

...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Se connecter et obtenir un token JWT
      description: |
        Authentifie un utilisateur avec son email et mot de passe.
        Retourne un token JWT à utiliser pour les requêtes authentifiées.
        
        **Compte de test:**
        - Email: `root`
        - Password: `root`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  description: |
                    Email de l'utilisateur. L'identifiant legacy `root` est également
                    accepté pour le compte d'administration de test.
                  example: "admin@thermosense.com"
                password:
                  type: string
                  format: password
                  minLength: 4
                  example: "root"
      responses:
        "200":
          description: Authentification réussie
          content:
            application/json:
              schema:
                type: object
                required: [token, user]
                properties:
                  token:
                    type: string
                    description: Token JWT à utiliser dans l'en-tête Authorization
                  user:
                    $ref: "#/components/schemas/User"
              example:
                token: "FAKE_JWT_TOKEN_example_user2_admin_do_not_use"
                user:
                  id: "user-2"
                  email: "admin@thermosense.com"
                  name: "Alice Admin"
                  role: "admin"
        "400":
          description: Payload invalide
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Email ou mot de passe incorrect
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /areas:
    get:
      tags: [Areas]
      operationId: listAreas
      summary: Lister toutes les zones
      description: |
        Retourne la liste paginée des zones enregistrées dans la plateforme.
        **Requiert une authentification.**

        **Roles autorises:** `admin`, `operator`, `reader`

        - `admin` : acces a toutes les zones
        - `operator`, `reader` : acces limite a leur zone assignee
      security:
        - BearerAuth: []
      x-required-roles: [admin, operator, reader]
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Liste paginée des zones.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AreasListResponse"
              example:
                data:
                  - id: "area-1"
                    name: "Zone A — Entrepôt Nord"
                    buildingId: "building-1"
                    sensors:
                      - id: "sensor-1"
                        type: "temperature"
                        status: "active"
                        areaId: "area-1"
                pagination:
                  limit: 50
                  offset: 0
                  total: 2
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags: [Areas]
      operationId: createArea
      summary: Créer une zone dans un bâtiment.
      description: |
        Créer une zone dans un bâtiment spécifique. La zone doit être associée à un
        bâtiment existant.
        **Requiert une authentification.**

        **Roles autorises:** `admin`
      security:
        - BearerAuth: []
      x-required-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [buildingId, name]
              properties:
                buildingId:
                  type: string
                  pattern: "^building-[0-9]+$"
                  description: Identifiant du bâtiment (format `building-{n}`).
                name:
                  type: string
                  minLength: 1
                  maxLength: 120
                  description: Nom de la zone
            example:
              buildingId: "building-1"
              name: "Zone C — Atelier Est"
      responses:
        "201":
          description: Création de la zone avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Areas"
              example:
                id: "area-3"
                name: "Zone C — Atelier Est"
                buildingId: "building-1"
                sensors: []
        "400":
          description: Payload invalide (buildingId et/ou name manquant)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /sensors:
    post:
      tags: [Sensors]
      operationId: createSensor
      summary: Créer un capteur dans une zone.
      description: |
        Enregistre un nouveau capteur et le rattache à une zone existante.
        **Requiert une authentification.**

        **Roles autorises:** `admin`, `operator`

        - `operator` : création limitée aux capteurs de sa propre zone
      security:
        - BearerAuth: []
      x-required-roles: [admin, operator]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, status, areaId]
              properties:
                type:
                  type: string
                  enum: [temperature, humidity]
                  description: Type de capteur.
                status:
                  type: string
                  enum: [active, inactive]
                  description: Statut opérationnel initial du capteur.
                areaId:
                  type: string
                  pattern: "^area-[0-9]+$"
                  description: Zone de rattachement (format `area-{n}`).
            example:
              type: "temperature"
              status: "active"
              areaId: "area-1"
      responses:
        "201":
          description: Capteur créé.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sensors"
              example:
                id: "sensor-8"
                type: "temperature"
                status: "active"
                areaId: "area-1"
        "400":
          description: Payload invalide (champ manquant ou valeur d'enum non autorisée).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: |
            Zone de rattachement introuvable. Pour un `operator`, une zone hors de son
            périmètre renvoie également `404` (masquage BOLA).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /users:
    post:
      tags: [Users]
      operationId: createUser
      summary: Créer un utilisateur
      description: |
        Crée un nouvel utilisateur avec l'email et le nom fournis.
        **Requiert une authentification.**

        **Roles autorises:** `admin`
      security:
        - BearerAuth: []
      x-required-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
            example:
              email: "operator.c@thermosense.com"
              name: "Camille Opératrice"
              role: "operator"
      responses:
        "201":
          description: Utilisateur créé.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              example:
                id: "user-8"
                email: "operator.c@thermosense.com"
                name: "Camille Opératrice"
                role: "operator"
        "400":
          description: Payload invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Email déjà utilisé.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  parameters:
...
    limit:
      name: limit
      in: query
      required: false
      description: Nombre maximal d'éléments retournés par page.
      schema:
        type: integer
        minimum: 1
        maximum: 499
        default: 50
    offset:
      name: offset
      in: query
      required: false
      description: Position de départ pour la pagination (0-indexé).
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    Unauthorized:
      description: |
        Token d'authentification absent, invalide ou expiré. Le client doit appeler
        `POST /auth/login` pour obtenir un nouveau token.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            type: "about:blank"
            title: "Unauthorized"
            status: 401
            detail: "Token JWT manquant ou expiré."
            code: "unauthorized"
    Forbidden:
      description: |
        Token valide mais rôle insuffisant pour cette opération, ou ressource
        appartenant à une autre zone (vérification BOLA).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            type: "about:blank"
            title: "Forbidden"
            status: 403
            detail: "Le rôle `reader` n'autorise pas la modification d'actionneurs."
            instance: "/actuators/actuator-1"
            code: "forbidden"
    TooManyRequests:
      description: |
        Quota d'appels dépassé. Voir la politique de rate limiting dans la description globale.
        Le header `Retry-After` indique le délai (en secondes) avant le prochain appel autorisé.
      headers:
        Retry-After:
          description: Délai en secondes avant un nouvel appel autorisé.
          schema:
            type: integer
            example: 30
        X-RateLimit-Limit:
          description: Quota total sur la fenêtre.
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Appels restants sur la fenêtre courante.
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Timestamp epoch (secondes) de remise à zéro du quota.
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            type: "https://thermosense.com/errors/rate-limit"
            title: "Too Many Requests"
            status: 429
            detail: "Quota de 120 req/min dépassé sur les endpoints de lecture."
            code: "rateLimitExceeded"
  schemas:
    Areas:
      type: object
      description: "Représente une zone dans un bâtiment."
      required:
        - id
        - name
        - buildingId
        - sensors
      properties:
        id:
          type: string
          pattern: "^area-[0-9]+$"
          description: "Identifiant unique de la zone (format `area-{n}`)."
          example: "area-1"
        name:
          type: string
          minLength: 1
          maxLength: 120
          description: "Nom de la zone."
          example: "Zone A — Entrepôt Nord"
        buildingId:
          type: string
          pattern: "^building-[0-9]+$"
          description: "Identifiant du bâtiment auquel la zone appartient (format `building-{n}`)."
          example: "building-1"
        sensors:
          type: array
          description: "Liste des capteurs présents dans la zone."
          items:
            $ref: "#/components/schemas/Sensors"
    Sensors:
      type: object
      description: "Représente un capteur dans une zone."
      required:
        - id
        - type
        - status
        - areaId
      properties:
        id:
          type: string
          pattern: "^sensor-[0-9]+$"
          description: "Identifiant unique du capteur (format `sensor-{n}`)."
          example: "sensor-1"
        type:
          type: string
          enum: [temperature, humidity]
          description: "Type de capteur."
          example: "temperature"
        status:
          type: string
          description: "Statut du capteur."
          enum: [active, inactive, maintenance]
          example: "active"
        areaId:
          type: string
          pattern: "^area-[0-9]+$"
          description: "Identifiant de la zone à laquelle le capteur appartient."
          example: "area-1"
...
    Error:
      type: object
      description: |
        Réponse d'erreur au format **Problem Details for HTTP APIs (RFC 7807)**.

        Les réponses d'erreur de l'API ThermoSense sont servies avec le media type
        `application/problem+json`. Les clients doivent traiter ce format quel que soit
        le code HTTP non-2xx retourné.

        **Extensions ThermoSense** (au-dessus de RFC 7807) :
        - `code` : identifiant machine-readable de l'erreur métier (`sensorUnavailable`,
          `idempotencyConflict`, `invalidParameter`, etc.) — stable, utilisable pour le routing
          côté client.
        - `errors` : liste des erreurs de validation par champ, présente sur les `400`/`422`.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
          description: |
            URI de référence identifiant le type d'erreur. Utiliser `about:blank`
            quand le code HTTP suffit à identifier la cause.
          example: "https://thermosense.com/errors/sensor-unavailable"
          default: "about:blank"
        title:
          type: string
          description: Résumé court et stable (ne change pas d'une occurrence à l'autre).
          example: "Sensor unavailable"
        status:
          type: integer
          description: Code HTTP, dupliqué ici pour faciliter le logging côté client.
          minimum: 400
          maximum: 599
          example: 409
        detail:
          type: string
          description: Explication contextuelle propre à cette occurrence.
          example: "Le capteur sensor-1 est hors-ligne depuis 2026-05-18T13:00:00Z."
        instance:
          type: string
          format: uri-reference
          description: URI identifiant l'occurrence (typiquement le chemin appelé).
          example: "/sensors/sensor-1/measures"
        code:
          type: string
          description: |
            Code d'erreur métier stable (extension ThermoSense).
            Exemples : `invalidParameter`, `sensorUnavailable`, `idempotencyConflict`,
            `idempotencyKeyMissing`, `rateLimitExceeded`, `forbidden`, `unauthorized`.
          example: "sensorUnavailable"
        errors:
          type: array
          description: |
            Erreurs de validation par champ (extension ThermoSense),
            présentes sur les réponses `400` et `422`.
          items:
            type: object
            required: [field, reason]
            properties:
              field:
                type: string
                description: Chemin du champ en erreur (notation pointée).
                example: "thresholdValue"
              reason:
                type: string
                description: Raison de l'erreur pour ce champ.
                example: "doit être strictement positif"
    PaginationMeta:
      type: object
      description: Métadonnées de pagination retournées avec toute liste.
      required: [limit, offset, total]
      properties:
        limit:
          type: integer
          description: Nombre maximal d'éléments demandé.
          example: 50
        offset:
          type: integer
          description: Position de départ utilisée pour cette page.
          example: 0
        total:
          type: integer
          description: Nombre total d'éléments disponibles côté serveur.
          example: 1440
    AreasListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Areas"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
...
    User:
      type: object
      description: "Utilisateur de la plateforme ThermoSense."
      required: [id, email, role]
      properties:
        id:
          type: string
          pattern: "^user-[0-9]+$"
          description: Identifiant unique de l'utilisateur (format `user-{n}`).
          example: "user-2"
        email:
          type: string
          description: |
            Email de l'utilisateur. Pour les comptes créés via `POST /users`,
            le format email standard est validé. Le compte d'administration
            historique `root` est conservé pour compatibilité.
          example: "admin@thermosense.com"
        name:
          type: string
          description: Nom affiché.
          example: "Alice Admin"
        role:
          type: string
          enum: [admin, operator, reader, device]
          description: Rôle applicatif attribué à l'utilisateur.
    UserCreateRequest:
      type: object
      required: [email, name]
      properties:
        email:
          type: string
          format: email
          description: Email de l'utilisateur.
        name:
          type: string
          minLength: 1
          maxLength: 120
          description: Nom affiché.
        role:
          type: string
          enum: [admin, operator, reader, device]
          description: Rôle applicatif (par défaut `reader`).
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Authentification JWT. Utilisez le token obtenu via `/auth/login`.
        
        **Pour tester:**
        1. Appelez POST `/auth/login` avec `{"email": "root", "password": "root"}`
        2. Copiez le token retourné
        3. Cliquez sur le bouton "Authorize" 🔒 en haut
        4. Entrez le token (sans le préfixe "Bearer")
        5. Toutes vos requêtes utiliseront automatiquement ce token !